wifi enable/disable ready. Big refactor for settings, wifi, variables on separated files
