--- src/main.hpp
@@ -1,1 +1,0 @@
-void wifiRestart();
