+++ src/wifi.h
@@ -1,0 +1,57 @@
+/******************************************************************************
+*   W I F I   M E T H O D S
+******************************************************************************/
+
+bool wifiCheck(){
+  wifiOn = WiFi.isConnected();
+  if(wifiOn)statusOn(bit_wan);  // TODO: We need validate internet connection
+  else {
+    statusOff(bit_cloud);
+    statusOff(bit_wan);
+  }
+  return wifiOn;
+}
+
+void wifiConnect(const char* ssid, const char* pass) {
+  Serial.print("-->[WIFI] Connecting to "); Serial.print(ssid);
+  WiFi.begin(ssid, pass);
+  int wifi_retry = 0;
+  while (WiFi.status() != WL_CONNECTED && wifi_retry++ < WIFI_RETRY_CONNECTION) {
+    Serial.print(".");
+    delay(250);
+  }
+  if(wifiCheck()){
+    isNewWifi=false;  // flag for config via BLE
+    Serial.println("done\n-->[WIFI] connected!");
+  }
+  else{
+    Serial.println("fail!\n-->[E][WIFI] disconnected!");
+    setErrorCode(ecode_wifi_fail);
+  }
+}
+
+void wifiInit(){
+  if(wifiEnable && ssid.length() > 0 && pass.length() > 0) {
+    wifiConnect(ssid.c_str(), pass.c_str());
+  }
+}
+
+void wifiStop(){
+  if(wifiOn){
+    Serial.println("-->[WIFI] Disconnecting..");
+    while(!WiFi.disconnect(true,true));
+    wifiCheck();
+  }
+}
+
+void wifiRestart(){
+  wifiStop();
+  wifiInit();
+}
+
+void wifiLoop(){
+  if(v25.size()==0 && wifiEnable && ssid.length()>0 && !wifiCheck()) {
+    wifiConnect(ssid.c_str(), pass.c_str());
+    influxDbReconnect();
+  }
+}
